requirements.txt
inflection
email-validator
fastapi[standard]
fastapi-pagination
isort
httpx
mimesis
psycopg2-binary
pycodestyle
pylint
pydantic
pytest
python-dotenv
sqlalchemy
sqlmodel
requests
uvicorn